pyproject.toml
[tool.poetry]
name = "MoViewer"
version = "0.1.0"
description = ""
authors = ["ahzees <matviy.stets@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
pillow = "^9.4.0"
pre-commit = "^3.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
